Upgrade Kotlin to 2.3.10 and release Phosphor 0.2.3 (#7)
* Upgrade Kotlin 2.1.20 to 2.3.10 and release Phosphor 0.2.3

- Update Kotlin version to 2.3.10 to match Ampere
- Bump Phosphor version to 0.2.3
- Add @OptIn(ExperimentalWasmDsl::class) for Kotlin 2.3.10 compatibility
- Remove redundant toFloat() calls in AgentLayer float calculations
- Fix JS float precision issues in substrate tests with tolerance-based comparisons
- Update yarn lock for new Kotlin version


* Add wasm yarn.lock generated by Kotlin 2.3.10 upgrade


---------

+++ phosphor-core/src/commonMain/kotlin/link/socket/phosphor/choreography/AgentLayer.kt
@@ -278,9 +278,9 @@ class AgentLayer(
             val phi = acos(1f - 2f * (i + 0.5f) / count)
 
             Vector3(
-                (sin(phi) * cos(theta) * radius).toFloat(),
-                (cos(phi) * radius).toFloat(),
-                (sin(phi) * sin(theta) * radius).toFloat(),
+                sin(phi) * cos(theta) * radius,
+                cos(phi) * radius,
+                sin(phi) * sin(theta) * radius,
             )
         }
     }
